PMM-7616 use TabbedContent on Settings
