+++ public/app/percona/settings/Settings.constants.ts
@@ -1,5 +1,45 @@
-export const GUI_DOC_URL =
-  'https://www.percona.com/doc/percona-monitoring-and-management/2.x/manage/server-admin-gui.html';
-export const DATA_RETENTION_URL =
-  'https://www.percona.com/doc/percona-monitoring-and-management/2.x/faq.html#how-to-control-data-retention-for-pmm';
-export const HOW_TO_DOC_URL = 'https://www.percona.com/doc/percona-monitoring-and-management/2.x/how-to';
+import { PageModel } from 'app/core/components/Breadcrumb';
+import { Messages } from './Settings.messages';
+import { TabKeys } from './Settings.types';
+
+export const DEFAULT_TAB = TabKeys.metrics;
+
+export const PAGE_TABS = [
+  {
+    title: Messages.tabs.metrics,
+    id: TabKeys.metrics,
+    path: `settings/${TabKeys.metrics}`,
+  },
+  {
+    title: Messages.tabs.advanced,
+    id: TabKeys.advanced,
+    path: `settings/${TabKeys.advanced}`,
+  },
+  {
+    title: Messages.tabs.ssh,
+    id: TabKeys.ssh,
+    path: `settings/${TabKeys.ssh}`,
+  },
+  {
+    title: Messages.tabs.alertManager,
+    id: TabKeys.alertManager,
+    path: `settings/${TabKeys.alertManager}`,
+  },
+  {
+    title: Messages.tabs.perconaPlatform,
+    id: TabKeys.perconaPlatform,
+    path: `settings/${TabKeys.perconaPlatform}`,
+  },
+  {
+    title: Messages.tabs.communication,
+    id: TabKeys.communication,
+    path: `settings/${TabKeys.communication}`,
+  },
+];
+
+export const PAGE_MODEL: PageModel = {
+  title: 'Settings',
+  path: 'settings',
+  id: 'settings',
+  children: PAGE_TABS.map(({ title, id, path }) => ({ title, id, path })),
+};
