+++ public/app/percona/settings/Settings.messages.ts
@@ -1,4 +1,8 @@
-import { GUI_DOC_URL, HOW_TO_DOC_URL } from './Settings.constants';
+export const GUI_DOC_URL =
+  'https://www.percona.com/doc/percona-monitoring-and-management/2.x/manage/server-admin-gui.html';
+export const DATA_RETENTION_URL =
+  'https://www.percona.com/doc/percona-monitoring-and-management/2.x/faq.html#how-to-control-data-retention-for-pmm';
+export const HOW_TO_DOC_URL = 'https://www.percona.com/doc/percona-monitoring-and-management/2.x/how-to';
 
 export const Messages = {
   advanced: {
